Changes to index.tsx to implement application toolbar
