+++ index.html
@@ -11,6 +11,6 @@
     <div id="adaptableBlotter"></div>
     <div id="grid" className="ag-theme-balham"></div>
 
-    <script src="./index.ts"></script>
+    <script src="./index.tsx"></script>
   </body>
 </html>
